Added small improvements
