+++ .gitignore
@@ -1,0 +1,2 @@
+flink-processor/target
+kafka-producer/target
