+++ kafka-producer/src/main/resources/log4j2.properties
